requirements.txt
# argus/requirements.txt

requests
python-dotenv